New upload
A ton of stuff that will add to the game.  When sorcerer or dragon
fights try to load, there is a stack overflow.  There is also another
known error early in the code that was meant to temporarily break it so
that another function could do its job printing stuff out.